image_tools/requirements.txt
onnxruntime-gpu==1.19.2
rembg==2.0.61
pillow==11.0.0